Move Conditions to persistence module.
