+++ config/src/main/java/com/alibaba/nacos/config/server/service/repository/extrnal/ExternalCommonPersistServiceImpl.java
@@ -16,7 +16,7 @@
 
 package com.alibaba.nacos.config.server.service.repository.extrnal;
 
-import com.alibaba.nacos.config.server.configuration.ConditionOnExternalStorage;
+import com.alibaba.nacos.persistence.configuration.condition.ConditionOnExternalStorage;
 import com.alibaba.nacos.config.server.constant.Constants;
 import com.alibaba.nacos.config.server.model.TenantInfo;
 import com.alibaba.nacos.persistence.datasource.DataSourceService;
