+++ persistence/src/main/java/com/alibaba/nacos/persistence/configuration/condition/ConditionOnExternalStorage.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,23 +14,23 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.config.server.configuration;
+package com.alibaba.nacos.persistence.configuration.condition;
 
 import com.alibaba.nacos.persistence.configuration.DatasourceConfiguration;
-import com.alibaba.nacos.sys.env.EnvUtil;
 import org.springframework.context.annotation.Condition;
 import org.springframework.context.annotation.ConditionContext;
 import org.springframework.core.type.AnnotatedTypeMetadata;
 
 /**
- * when embeddedStorage==true and nacos.standalone=false
+ * Judge whether to user ExternalStorage by condition.
  *
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
-public class ConditionDistributedEmbedStorage implements Condition {
+public class ConditionOnExternalStorage implements Condition {
     
     @Override
     public boolean matches(ConditionContext context, AnnotatedTypeMetadata metadata) {
-        return DatasourceConfiguration.isEmbeddedStorage() && !EnvUtil.getStandaloneMode();
+        return !DatasourceConfiguration.isEmbeddedStorage();
     }
+    
 }
