+++ persistence/src/test/java/com/alibaba/nacos/persistence/configuration/condition/ConditionDistributedEmbedStorageTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.config.server.configuration;
+package com.alibaba.nacos.persistence.configuration.condition;
 
-import com.alibaba.nacos.config.server.utils.PropertyUtil;
 import com.alibaba.nacos.persistence.configuration.DatasourceConfiguration;
 import com.alibaba.nacos.sys.env.EnvUtil;
 import org.junit.Assert;
@@ -41,18 +40,19 @@ public class ConditionDistributedEmbedStorageTest {
     @Before
     public void init() {
         conditionDistributedEmbedStorage = new ConditionDistributedEmbedStorage();
-       
+        
     }
     
     @Test
     public void testMatches() {
-        MockedStatic<PropertyUtil> propertyUtilMockedStatic = Mockito.mockStatic(PropertyUtil.class);
+        MockedStatic<DatasourceConfiguration> propertyUtilMockedStatic = Mockito
+                .mockStatic(DatasourceConfiguration.class);
         MockedStatic<EnvUtil> envUtilMockedStatic = Mockito.mockStatic(EnvUtil.class);
         
         propertyUtilMockedStatic.when(DatasourceConfiguration::isEmbeddedStorage).thenReturn(true);
         envUtilMockedStatic.when(EnvUtil::getStandaloneMode).thenReturn(true);
         Assert.assertFalse(conditionDistributedEmbedStorage.matches(context, metadata));
-    
+        
         Mockito.when(DatasourceConfiguration.isEmbeddedStorage()).thenReturn(true);
         Mockito.when(EnvUtil.getStandaloneMode()).thenReturn(false);
         propertyUtilMockedStatic.when(DatasourceConfiguration::isEmbeddedStorage).thenReturn(true);
